requirements.txt
mcp
websockets
python-dotenv
